Created PhoenixPlayer object

+++ PhoenixRevival/src/main/java/com/github/elfin8er/PhoenixRevival/SETTINGS.java
@@ -2,7 +2,7 @@ package com.github.elfin8er.PhoenixRevival;
 
 public class SETTINGS {
 
-	public final static int StartingMoney = 100;
+	public final static Double StartingMoney = 100.00;
 	public final static String CurrencySymbol = "₱";
 	
 }
